Turn incorrect usage (use-after-close) from SEGFAULT into clean errors in most cases.
See the comment added to the `DB` type for details.

Segfaults are very nasty to track down and don't report an error message
that end users can report to developers, so any case where we can turn
a segfault into a clean error message is an improvement.

Also add `withDB` function for clean scoped opening, and add `HasCallStack`
constraints on functions that can error so that callers can more easily spot
where in their code the mistake is if we can spot it.

This also fixes the tests again that were segfaulting as introduced
in the previous commit.